compass tracks shops & bugfixes

+++ items/compass.cs
@@ -13,11 +13,61 @@ datablock PlayerData(MiniDungeonsCompassArmor : PlayerStandardArmor)  {
 	uiName = "";
 };
 
+datablock ItemData(CompassItem) {
+	category = "Weapon";
+	className = "Weapon";
+
+	shapeFile = "./shapes/compass.dts";
+	rotate = false;
+	mass = 1;
+	density = 0.2;
+	elasticity = 0.2;
+	friction = 0.6;
+	emap = true;
+
+	uiName = "Compass";
+	iconName = "./";
+	doColorShift = false;
+
+	image = CompassImage;
+	canDrop = true;
+};
+
+datablock ShapeBaseImageData(CompassImage) {
+	shapeFile = "base/data/shapes/empty.dts";
+	emap = true;
+
+	mountPoint = 0;
+	offset = "0 0 0";
+	correctMuzzleVector = true;
+
+	className = "WeaponImage";
+
+	item = HealthFlaskItem;
+	ammo = " ";
+
+	melee = false;
+	armReady = true;
+
+	doColorShift = false;
+
+	interactablePassthrough = true;
+};
+
+function CompassImage::onMount(%this, %obj, %slot) {
+	Parent::onMount(%this, %obj, %slot);
+	%obj.showCompass();
+}
+
+function CompassImage::onUnMount(%this, %obj, %slot) {
+	Parent::onUnMount(%this, %obj, %slot);
+	%obj.hideCompass();
+}
+
 function Player::showCompass(%this) {
+	%this.createSwordHands();
 	%this.swordHands.unHideNode("rhand");
-	for(%i = 0; %i < 4; %i++) {
-		%this.swordHands.playThread(%i, "holdCompass");	
-	}
+	%this.swordHands.playThread(0, "holdCompass");
 	%this.playThread(1, "armReadyRight");
 
 	if(!isObject(%this.compass)) {
@@ -32,15 +82,27 @@ function Player::showCompass(%this) {
 
 	%this.hideRightHand = true;
 	%this.client.applyBodyParts();
+
+	%shop = 0;
+	if(isObject(%room = %this.getCurrentRoom())) {
+		%shop = %room.roomFindClosestType(1);
+	}
+	else if(isObject(%hallway = %this.getCurrentHallway())) {
+		%shop = %hallway.room1.roomFindClosestType(1);
+	}
+
+	if(isObject(%shop)) {
+		%this.compassTrackPosition(%shop.worldPosition);
+	}
 }
 
 function Player::hideCompass(%this) {
-	%this.hideRightHand = false;
-	for(%i = 0; %i < 4; %i++) {
-		%this.swordHands.playThread(%i, "root");	
+	if(isObject(%this.compass)) {
+		%this.hideRightHand = false;
+		%this.swordHands.playThread(0, "root");
+		%this.compass.delete();
+		%this.client.applyBodyParts();
 	}
-	%this.compass.delete();
-	%this.client.applyBodyParts();
 }
 
 function Player::compassTrackPosition(%this, %position) {
